raspberry-pi/requirements.txt
google-cloud-pubsub
google-cloud-storage
python-dotenv
smbus